Moved character creation to it's own .dart file
